chore: apply separated components
